// src/app/types/enums.ts
export enum AccountCategory {
  REVENUE = "revenue",
  EXPENSE = "expense",
  ASSETS = "assets",
  LIABILITY = "liability",
}

export enum ValueType {
  CREDIT = "credit",
  DEBIT = "debit",
}

export enum AccountType {
  SALES = "sales",
  OVERHEADS = "overheads",
  CURRENT_AR = "current_accounts_receivable",
  BANK = "bank",
  FIXED = "fixed",
  CURRENT_AP = "current_accounts_payable",
  TAX = "tax",
  CURRENT = "current",
  PAYROLL = "payroll",
}